DCON-3672: Expand auth-failure test coverage across endpoint groups
Move the test file from src/tests/graphqlv2/auth_failure/ to
src/tests/auth_failure/ and add one test per endpoint group: REST CRUD,
$merge, $everything, $export, graphql, graphqlv2. Each asserts the
shared OperationOutcome JSON body on missing auth.

$export requires ENABLE_BULK_EXPORT='1' at module load so the route is
registered before the first app construction in this file.
